Stub TypeDecls for Types from Header Files
For C/C++ types from header files without any accessible implementation
we might end up with CPG Types without a corresponding CPG TypeDecl.
As these types may be referenced via `TypeDecl.aliasTypeFullName` e.g., for C++ type alias or typedefs we previously crashed e.g., in `TypeDeclTraversal.canonicalType`.

This PR stubbs them.

Fixes: https://github.com/ShiftLeftSecurity/codescience/issues/5784
