+++ joern-cli/frontends/c2cpg/src/main/scala/io/joern/c2cpg/C2Cpg.scala
@@ -1,7 +1,7 @@
 package io.joern.c2cpg
 
 import io.joern.c2cpg.C2Cpg.Config
-import io.joern.c2cpg.passes.{AstCreationPass, HeaderContentLinkerPass, PreprocessorPass}
+import io.joern.c2cpg.passes.{AstCreationPass, HeaderContentPass, PreprocessorPass}
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.Languages
 import io.shiftleft.passes.{IntervalKeyPool, KeyPoolCreator}
@@ -16,11 +16,12 @@ import scala.util.control.NonFatal
 class C2Cpg {
 
   def runAndOutput(config: Config): Cpg = {
-    val keyPool = KeyPoolCreator.obtain(3, minValue = 101)
+    val keyPool = KeyPoolCreator.obtain(4, minValue = 101)
     val metaDataKeyPool = new IntervalKeyPool(1, 100)
     val typesKeyPool = keyPool.head
     val astKeyPool = keyPool(1)
     val headerKeyPool = keyPool(2)
+    val headerContentKeyPool = keyPool(2)
 
     val cpg = newEmptyCpg(Some(config.outputPath))
 
@@ -31,10 +32,10 @@ class C2Cpg {
     val headerAstCreationPass = new AstCreationPass(cpg, AstCreationPass.HeaderFiles, Some(headerKeyPool), config)
     headerAstCreationPass.createAndApply()
 
-    new HeaderContentLinkerPass(cpg, config).createAndApply()
-
     val types = astCreationPass.usedTypes() ++ headerAstCreationPass.usedTypes()
     new TypeNodePass(types.distinct, cpg, Some(typesKeyPool)).createAndApply()
+
+    new HeaderContentPass(cpg, Some(headerContentKeyPool), config).createAndApply()
     cpg
   }
